+++ PrivateChatClient/PrivateChatClientDlg.h
@@ -40,4 +40,5 @@ public:
 	afx_msg void OnBnClickedButton2();
 	afx_msg void OnTimer(UINT_PTR nIDEvent);
 	afx_msg void OnBnClickedButton3();
+	CEdit m_EE2EToken;
 };